requirements.txt
fastapi[standard]
python-dotenv
pyjwt
bcrypt
SQLAlchemy
psycopg2